% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdplyr-package.R
\docType{package}
\name{sdplyr-package}
\alias{sdplyr}
\alias{sdplyr-package}
\title{sdplyr: Apache SedonaDB 'dplyr' Interface}
\description{
Provides a 'dplyr' backend implemented using 'SedonaDB'.
}
\author{
\strong{Maintainer}: Dewey Dunnington \email{dewey@dunnington.ca}

Authors:
\itemize{
  \item Dewey Dunnington \email{dewey@dunnington.ca}
}

}
\keyword{internal}
